fix: prosemirror-view patch
