+++ prosemirror-jdita/package.json
@@ -1,11 +1,11 @@
 {
   "name": "prosemirror-jdita",
-  "version": "0.0.3",
+  "version": "0.0.5",
   "description": "Use JDita in prosemirror",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
   "scripts": {
-    "fix": "node fix",
+    "fix": "node scripts/fix",
     "postinstall": "yarn fix",
     "lint": "eslint src --ext .ts",
     "lint:fix": "eslint src --ext .ts --fix",
@@ -33,11 +33,11 @@
   "homepage": "https://github.com/ccheraa/prosemirror-jdita#readme",
   "directories": {
     "src": "./src",
-    "test": "./test"
+    "test": "./test",
+    "scripts": "./scripts"
   },
   "files": [
-    "lib",
-    "fix.js"
+    "lib"
   ],
   "dependencies": {
     "@types/chai-as-promised": "^7.1.3",
